server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deepdiff==6.7.1
fastapi==0.110.2
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
ordered-set==4.1.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pylitterbot==2023.5.0
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0
yarl==1.9.4
